% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\name{add}
\alias{add}
\title{Add Two Numbers}
\usage{
add(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
a numeric value
}
\description{
A function to add two numbers
}
\author{
Thomas Rosendal
}

