% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Addition_Function.R
\name{additionloop}
\alias{additionloop}
\title{Adds consecutive numbers together
 
 @param x A numerical vector
 @return The sum of all numbers in the vector
 @examples additionloop(1:10)
 @export}
\usage{
additionloop(x)
}
\description{
Adds consecutive numbers together
 
 @param x A numerical vector
 @return The sum of all numbers in the vector
 @examples additionloop(1:10)
 @export
}
