% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sampledata}
\alias{sampledata}
\title{sampledata}
\usage{
sampledata()
}
\value{
a dataframe
}
\description{
Some data that is really interesting
}
\details{
Read in the sample data
}
\author{
Thomas Rosendal
}

